First try at the bowling game
I think I have the normal game logic down. Right now it won't work if
you want to be able to print every frame mid game. So the next steps
would be to print running totals of the score as the game is happening
instead of just the total at the end of the game.